Improve unification + solving
Instead of immediately failing when we encounter a repeated instance, we
push an equality constraint to the _end_ of the stack. This will
hopefully allow the solver to perform more substitutions before deciding
that two types can't be unified.

Instance constraints now result in substituting the variable over the
set of constraints. For instance (pardon the pun), if we were solving
the following set of constraints

   'a ~ 'b
   'b ~ int
   'c ~ 'a -> int

The first iteration of the solver would eliminate 'a from the set
_entirely_:

  'b ~ int
  'c ~ 'b -> int

We then eliminate 'b:

  'c ~ int -> int

And end up with a trivial constraint.