Fix(??) GADT-induced unsoundness
We do this by adding a new field to `Var Typed`, that indicates if the
variable is rigid (skolem) or flexible. If it *is* rigid, then we can
not instance/bind that type variable: We reject this both in
`instantiate` and `unify`. This prevents counterexamples/skolems.ml from
compiling. Success!

Rigid variables are those introduced by a GADT constructor and not
mentioned in the header, i.e
  type hide =
    Hide : 'a -> hide