Remove peval
While a great idea in theory, practice is almost never the way we want
it to be. What we'd like to do is reduce as much code as possible at
compile time so that runtime performance is improved. What we end up
doing is spend significantly more time in the compiler to end up doing
what would be cheap operations at runtime anyway.

For instance, eliminating conversion between native integers and
Peano-style inductive naturals is basically impossible for peval, as it
is implemented now, to perform on anything other than a toy scale
(folding `from_int 10` takes 93959 reduction steps, and 1.12 seconds,
while the corresponding unoptimised program runs almost instantly).

I don't plan on removing all optimisations - I just plan on not doing
them all in one go - more modularity is always good, and in this case
especially verifying correctness of several small passes is a lot
simpler than verifying correctness of a single big pass.

You might have noticed this is in a new branch - I don't feel like this
is a branch that'll be merged once. Instead, I think that most optimiser
improvements will happen here and be merged incrementally into master.