Use optimiser passes (again) (#1)
* Remove peval

While a great idea in theory, practice is almost never the way we want
it to be. What we'd like to do is reduce as much code as possible at
compile time so that runtime performance is improved. What we end up
doing is spend significantly more time in the compiler to end up doing
what would be cheap operations at runtime anyway.

For instance, eliminating conversion between native integers and
Peano-style inductive naturals is basically impossible for peval, as it
is implemented now, to perform on anything other than a toy scale
(folding `from_int 10` takes 93959 reduction steps, and 1.12 seconds,
while the corresponding unoptimised program runs almost instantly).

I don't plan on removing all optimisations - I just plan on not doing
them all in one go - more modularity is always good, and in this case
especially verifying correctness of several small passes is a lot
simpler than verifying correctness of a single big pass.

You might have noticed this is in a new branch - I don't feel like this
is a branch that'll be merged once. Instead, I think that most optimiser
improvements will happen here and be merged incrementally into master.

* Now with less points!

* Who the hell derives Read?

* Implement match-of-known-constructor

* boilerplate

* Add trivial & constructor-application propagation

Trivial propagation moves literals to use sites instead of allocating a
variable.

Constructor-application propagation does the following transformation:
  let { v = C x } in ... v ...
=>
  let { v' = x } in ... C v ...

The rationale for this is that, by exposing constructor applications at
use sites, we get more opportunities to do match-of-known-constructor.

* Optimise more times

@SquidDev could you make the optimiser do things until a fixed point is
reached?

* Export fresh

* Revert "Merge branch 'master' into feature/optimiser-passes"

This reverts commit 6e556d18fb6f31b33bfa50a6c8cca7a3c3150961, reversing
changes made to 941e69f3e85e9972f49f99268976136d646458b8.

* save me from this madness

* you know the gist

* Kill before/after distinction

* Record number of changes done during a round of optimisation

* Add an inliner

Inlining is currently done in two phases:
  - inlining itself, which turns a variable into its definition if the
  definition is cheap enough
  - beta-reduction, which removes applications of lambdas to terms.

* Don't emit trailing commas in tables

Otherwise we end up with { a,  }, which just looks ugly.