Several type-checking bugfixes.
Apparently: The order in which you resolve constraints is important, so
sorting them is bogus. That's OK.

Less OK: The order of parameters to `unify` matters. This made
application checking be broken for ages, but I fixed it now. It's
mind-boggling.

Even less OK: We don't have any automated testing to prevent this from
happening. @SquidDev any ideas here?