Modules (#5)
Squashes are for hiding the embarrassing mistakes, right?
Well, here they are.

* Implemented scope extraction

* Implement parsing of qualified names and modules

* A primitive module implementation

Open just generates an error message currently.

* lensify

* lensify better

* Use Semigroup instead of custom type class

* Some improvements to printing

* Straighten out records

Now they're less of a hack. I like this. Sorry for doing this on the
wrong branch, Squid.

* Infer `open`

* Improve modules example

This is a more general example of the various features.

* Add the worst module resolver

I've no clue if this works, or if it works with any of the type inference
stuff. Welp.

* The less debug traces, the better

* Open shouldn't define any exportable variables

* Polish up

* shit I'm dumb

* Lookup in parent modules if possible

* Simplify module handling of the inferer