Re-write inferLetTy
For starters, we're back to inferring the types of bindings, so the
function isn't totally misnamed.

It now works on individual SCCs of the graph at a time, which makes
things a lot easier in the single variable case (or, well, even when
they are several variables and they are not mutually recursive). We make
up type variables as needed, and unify them with the result of
inference.

Solving is done once, after inference, and all of the variables get the
solution applied to them in one go.