Use types based on programmer annotations
Instead of blindly making up fresh type variables for `let`, we can use
the fact that Amulet has lots of possibilities to sneak ascriptions in
to let the programmer supply - either completely or partially - a type
signature.

Inference for lets now begins by approximating the type, and marking it
as either "guessed" (some parts of it are fresh type variables made up
for this) or "supplied" (the programmer completely specified this type).

You might wonder what gains this brings us if we have complete type
inference - but the problem is, we do *not* have complete type
inference! Namely, our system allows polymorphic recursion, but the type
checker will never infer it.

However, if the programmer completely supplies a type signature, then we
can trust them (to some extent!):

  let length (x : nested 'a) : int =
    match x with
    | Epsilon -> 0
    | Sub (_, xs) -> 1 + length xs ;;

  let length : forall 'a. nested 'a -> int = fun x ->
    match x with
    | Epsilon -> 0
    | Sub (_, xs) -> 1 + length xs ;;

Unfortunately, the type checker isn't quite enough to do partial
matching. For instance, the definition below is rejected:

  let length (x : nested 'a) =
    match x with
    | Epsilon -> 0
    | Sub (_, xs) -> 1 + length xs ;;

This is because we made the compiler guess the codomain of `length`.