Fever-induced Generalised Algebraic Data Types
I decided to see if I could do this in under an hour. I took 73 minutes.
Calling that a success.

This adds a new form of constructor, the "generalised" constructor:

  | Foo : t

`t` must "end" in an application of the type being declared, i.e.

    forall vs. t where t is kosher
    a -> t where t is kosher
    an application of the type being declared

Pattern matching on a GADT constructor introduces local equalities /when
checking/. This lets us guarantee principal types, I think. The details
are fuzzy.

Matching on a GeneralisedCon that binds constraints 'cs' will wrap the
body's generated constraints in an implication:

  type term 'a =
    | Lit : 'a ~ int => int -> term 'a
    | Unit : 'a ~ unit => term 'a
    | Fun : forall 'c 'b. 'a ~ 'c -> 'b => ('c -> 'b) -> term 'a
    | App : forall 'c 'b. 'a ~ 'b => term ('c -> 'b) * term 'c -> term 'a
  ;;

  let eval (x : term 'a) : 'a =
    match x with
    | Lit l -> l

This is ok because the constraints end up looking like
  a ~ int => a ~ int
which is trivial. type variables mentioned in the left-hand-side of an
implication constraint (the "before") are not kept in the resulting
substitution.