This is a lint failure now
I'm strangely OK with this. Using the FFI to convince the compiler two
concrete types are equal is bogus, especially since there are no surface
level guarantees that two types are ever the same.

Since we don't have roles (all coercions are nominal, which means, ehm,
core is very unsound), by using transitivity of equality presented in
this example one could derive `s ~ k` starting from `forall 'a. s 'a ~ k
'a` (very easily), which is a bit bogus.

The optimiser, then, gets confused with this (which makes a surprising
amount of sense), and then lint bails.