Add an example of stream fusion
This shows off several things:

- Join point optimisations let fusion actually take place, as can be
seen by inspecting the generated code. Any mention of Stream/the step
constructors is entirely elided from the loop, which is pretty much as
tight as it could be (except for boxing/unboxing the tuple);

- The support for stream transformers, producers, and consumers,
independent of inner state, relies crucially on existentials working
properly; and

- Other, more boring optimisations such as match β-reduction and
inlining.