Numerous fixes to the backend
This ensures that expressions are always emitted in-order, as well as
fixing several bugs with expressions being evaluated multiple times.

Sadly, as our expression ordering is not entirely optimal, we don't
always emit as good code as we should: see the arithemetic example for
instance. Fixing this will require a more comprehensive "reordering"
pass.