Componentwise Tuple Subsumption (#61)
This adds Amulet's first two wired-in names with unfoldings, 'fst#'
(-32) and 'snd#' (-33). The GHC-inspired magic hash is there to indicate
magic; These bindings are not user accessible and are basically only
added by the type checker.

'fst#' and 'snd#' are used to re-build tuples at use sites that impose
subsumption constraints. Consider we need to solve

    "('a * 'b) <= ('c * 'd)"

We start by decomposing these into

    'c <= 'a (call this wrapper α)
    'd <= 'b (call this wrapper β)

Then, we re-build the tuple (call it x, with ξ fresh) like so:

    let ξ = x (* to prevent reevaluation *) in
      (α[fst ξ], β[fst ξ])

where w[e] means applying w to e.

This also fixes a long-standing bug where some sequences of wrappers 
were generated backwards. This was basically impossible to hit before.