Allow explicit impredicativity
Previously, we banned any use of a polytype under a type constructor.
Now, all of these uses have to be annotated. So, for instance, a list of
identity functions could be written

  type list 'a =
    | Nil : forall ('a : type) -> list 'a
    | Cons : forall ('a : type) -> 'a * list 'a -> list 'a

and used like

  let ids = Cons @{forall 'a. 'a -> 'a}
    (fun x -> x, Nil @{forall 'a. 'a -> 'a})
  let head : forall ('a : type) -> list 'a -> 'a =
    fun (Cons (a, _)) -> a
  let main = head @{forall 'a. 'a -> 'a} ids 1

This is a tad crap, but it's probably good enough for now, in true
Amulet tradition.