WIP: REPL (#60)
This accepts top-level statements or expressions and evaluates them. These
are then fed through the compiler pipeline and evaluated using HSLua.

It's currently rather primitive in places (no autocompletion, no commands,
etc...) but it's a good framework and makes it easier to fiddle with things.

This also makes several improvements to pretty printing and the type
checker.