New-type elimination (again) (#67)
* Re-introduce the newtype pass

Newtype optimisation is back! Now with, hopefully, 100% less broken Core
Lints. I've added a small test about how newtypes are compiled (this
is likely to break in the future if we ever manage to inline
constructors, by the way).

This optimisation is unreasonably efficient in Amulet, considering that
all types with a single constructor (and no existentials) are eligible
for newtype elimination, considering we have no curried constructors.
This includes single-constructor GADTs, since evidence parameters are
currently erased during Lower.

The idea is the same as before: we turn a type like
  type foo 'a {
    Foo : forall ('a : *). 'a -> foo 'a
  }
into a worker like
  type foo 'a {}
  let Foo : forall ('a : *). 'a -> foo 'a =
    Λ ('a : *). λ(x : 'a). x |> 'a ~ foo 'a
which will probably inline everywhere.