Add deep subsumption
This rule was inspired by Martin Odersky and Konstantin Läufer's 1996
paper "Putting Type Annotations to Work".

In short: we can conclude
  σ₁ -> t <= σ₂ -> t
if we can conclude σ₂ <= σ₁. The actual implementation also allows 't'
to be distinct, to end up at the rule

  σ₂ <= σ₁ |- f                   τ₁ <= τ₂ |- g
  ---------------------------------------------
  σ₁ -> τ₁ <= σ₂ -> τ₂ |- \ex. λy. g (ex (f y))

using \ex to stand for a wrapper continuation, and |- reads "produces
the wrapper".