Compute more complete outline types
Consider this (admittedly contrived) program

  let size (x : nested 'a) =
    match x with
    | One _ -> { count = 1 }
    | Nested (a, _) -> { count = 2 * (size a).count }

Before, this would fail with the occurs check trying to
unify 'a and ( 'a * 'a ). Now, since we can determine

  { count = 1 } : { count : int }

before calling out to 'infer', and thus can concretely determine that

  size : forall 'a. nested 'a -> { count : int }

from looking at the declaration alone, we use the checking judgement
which allows polymorphic recursion.