Rewrite the Lua codegen
The previous version of the codegen operated on the concept of an
expression "stack". We'd push new variables onto the stack, and pop them
when they were consumed. If a variable did note match up, we'd convert
the entire stack into local variables, to ensure that the order of side
effects was preserved.

While this worked well for unprocessed inputs, the optimiser would often
reorder terms in such a way that the codegen would flush bindings more
often than needed, meaning we'd generate rather silly Lua.

This rewritten codegen handles execution order via a graph instead. We
walk down the spine of let expressions, and gather expressions and their
dependencies (free variables and the preceding impure expression).

Expressions which are only used once can be embedded inline as long as
doing so does not introduce a conflict (determined by a loop in the
dependency graph). Otherwise they are emitted as individual statements
and emitted to the graph.

When a block has been fully generated, we perform a topological sort of
the dependency graph, concatenating all statements together.

The other noticeable difference compared to the previous generator is
the introduction of "yield contexts". These simply determine how a given
node is consumed - it could be discarded, returned, declare a new
variable or bound to an existing one. This allows us to perform more
fine-grained code generation is several cases (such as when emitting
matches or record extensions).