Fix pattern typing-variable typing confusion
We need to somehow distinguish between the type of the bound variable
(what the rest of the program sees) and the type of what is matched (the
type of the pattern itself). Consider:

  let (x : t, ...) = (foo, ...)
   in ... x : ∀ a. t ...

Before, we would lower assuming that ∀ a. t was the type of *both*,
which led to endless woe.