Feature: Pattern Matching Lets (#81)
Now you can do

    let (a, b) = ...
    a + b

The generated core for this ends up being a little grim, as 
we need to handle polymorphic bindings and ensure the
generation works for both top-level and term-level lets.

I can't speak about how grim the type checker is, as I didn't
do much with it. The parser and resolver are pretty simple
though!