Allow substituting any constructor with a single possible value
This will reduce expressions of the form

  let x = EXPR in
  y

to

  EXPR

If x and y have compatible types and x is a type with a single possible
value.

Note this optimisation is not currently perfect, as it will not work for
polymorphic constructors. However, most single value constructors will
not be polymorphic [citation needed].