Reducer rewrite pt 2: Inliner Boogaloo (#101)
* Move the inliner into the reducer

This is probably slightly more inefficient (as we score nodes every time
and whatnot). Thankfully this isn't even its final form.

* Move some pattern code into a separate module

We should be able to utilise this within the inliner too.

* Keep track of free variables within the reducer

 - Split up recursive lets which are no longer recursive

 - Add loop-breaker analysis, and allow non-loop breaker variables to be
   inlined. This is still slightly incomplete, as we'd like to use usage
   information here too, but it is sufficient for now.

* Keep track of occurrence information as well

Oh boy, this is beginning to get ugly

* Hopefully be smarter about how single-use variables are inlined

* Track purity mildly better

* Be much more selective in what we defer resolving

We now only consider pure applications and lambdas, as they should be
picked up by the inliner. While it _could_ be beneficial to handle other
cases, this would end up being little more than a underwhelming sinking
pass.