Error when a stated type signature is insufficient
A *stated* type signature is one that the parser has shoved into a
toplevel ascription onto a value.
  let foo : ty = ...
  let foo = ... : ty

A *deduced* type signature, on the other hand, is something like this:
  let foo (x : 'a) = 1
  (* here we have deduced forall 'a. 'a -> int *)

The names were chosen to imply some deeper connection with type
inference. Indeed, the type checker could infer the second signature all
on its own: the fact we precompute it shouldn't matter at all. We're
free to modify this type signature, though, as long as none of the
stated type variables change.

So, indeed, we can add a context to deduced type signatures but not
stated type signatures.