Invert instance desugaring
Instead of making a single non-recursive dictionary binding that
contains a binding group with all the methods ($dUse_old), we make a
recursive binding group with the methods _and_ the dictionary
($use/$dUse_new).

>  let $dUse_old =
>    fun @a (cf : use br'a) -> let use : list br'a -> unit =
>      fun f -> match f with
>      | Nil -> ()
>      | (Cons (_, xs)) ->
>          use {list br'a} ($dUse_old @a cf) xs
>    in Use @(list br'a) use
>  let $use =
>    fun @a (cg as cf) f -> match f with
>    | Nil -> ()
>    | (Cons (_, xs)) ->
>        use {list br'a} ($dUse_new @a cg) xs
>  and $dUse_new : {'a : type}. use 'a => use (list 'a) =
>    fun @a (cg as cf : use br'a) -> Use @(list br'a) ($use @br'a cf)

This should mean that the optimiser is able to inline methods and
dictionaries, but it does result in generally better code anyway - See
e.g. tests/lua/default-method, which is now entirely obsolete because
the compiler sees through the lowered recursive dictionary and inlines
the method.