Refine VTA and (re-)add required 'forall's
The Amulet compiler now has a notion of type-variable binder
/specificity/: whether a type variable was inferred from the type or
specified by the user in a 'forall'.

Namely, there are three specificities: Req (forall 'a ->), Spec (forall
'a.) and Infer (forall {'a}., though this can't be parsed).

'forall' binders CAN ONLY be eliminated if they are Req or Spec. Req
binders MUST be eliminated visibly.

Variables are:

  - Spec if they are introduced by a user-written type variable binder:
  class parameters, type definition parameters, or a literal forall
  type.

  - Req if they are introduced by a type such as
    forall 'a -> 'a -> 'a

  - Infer if they arise solely during type checking.

This means that:

* The type of
    let id x = x
is
    forall { 'a }. 'a -> 'a
and the variable 'a is not subject to VTA;

* The type of
    let id (x : 'a) : 'a = x
is
    forall 'a. 'a -> 'a
and the variable 'a is subject to VTA;

* The type of
    let id : forall 'a -> 'a -> 'a = fun x -> x
is
    forall 'a -> 'a -> 'a
and the variable 'a MUST BE visibly applied.

* Kind variables introduced from class/type definitions can never be
visibly applied.