Somewhat update the context system to work with do notation
'let's are now handled correctly in the multi-line case. However, on the
single line case, `let a = 1; 2` is parsed as `let a = (1; 2)`, which is
technically correct but not desirable.

In a desperate desire to avoid possible confusion, we parse 'with blocks
in an expression block (say, let bodies), and then produce a custom
error instead.