Add monad syntax (#152)
* Add monad syntax

* Somewhat update the context system to work with do notation

'let's are now handled correctly in the multi-line case. However, on the
single line case, `let a = 1; 2` is parsed as `let a = (1; 2)`, which is
technically correct but not desirable.

In a desperate desire to avoid possible confusion, we parse 'with blocks
in an expression block (say, let bodies), and then produce a custom
error instead.

* Generate alex/happy files in a disjoint directory

I just spent 5 minutes wondering why my changes weren't applying - turns
out stack was using the pre-existing parser.