Start lowering and linting all TC tests
While this may seem a little overboard, type-checking tests (obviously)
have the largest coverage of the type system, and so it makes sense to
check all these features can be successfully lowered.

Note, this creates a lower test for /every/ TC test, irrespective of
whether it actually TCs or not. This is unideal, but the easiest thing
to do for now.