Make core lint opt-in
 - Core lint is now configured by an argument to simplify, rather than
   being a compile-time setting.
 - `amc compile' now accepts a --core-lint setting, which will enable
   linting.

Note, the REPL does not currently support this, and will never run core
lint right now.

Ideally this would be a setting within the driver, but lint currently
requires a full program to run. We would need a way to convert a
LowerState into a Lint.Scope in order to do this though.