Tiny module-system and CLI improvements (#206)
* Rename "modules" to "file" within the driver

This should have no impact on behaviour or the external API. Just trying
to disambiguate between a "compilation unit" (i.e. a file), and a
module. While every file is a module, the inverse is not true, and so we
try to use a more appropriate name.

* Make core lint opt-in

 - Core lint is now configured by an argument to simplify, rather than
   being a compile-time setting.
 - `amc compile' now accepts a --core-lint setting, which will enable
   linting.

Note, the REPL does not currently support this, and will never run core
lint right now.

Ideally this would be a setting within the driver, but lint currently
requires a full program to run. We would need a way to convert a
LowerState into a Lint.Scope in order to do this though.

* Move prelude handling into the driver

 - Adds a --prelude/--no-prelude flag to the compile and repl
   sub-commands. Note, this is dubiously safe, as it means libraries
   will also be loaded with this prelude.
 - compile will attempt to open the prelude for all compiled files.

* Do not load the prelude when compiling

This makes prelude-loading REPL specific /by choice/. While
prelude-by-default would be nice, it's not immediately clear how
external libraries should interact with a custom prelude. Furthermore,
the implementation is a little ugly, so we'll just remove it for now.

* Allow using --core-lint in amc repl and amc compile