Make core less polymorphic
We remove the "variable" type parameter from Type, Atom and Coercion, as
the information in these nodes was never correct/useful. This reduces
the amount of mapping between Type a and Type (Occurs a) within the Lua
backend and the reduce pass.

In the future, I would like to convert all optimisation passes to work
on a concrete CoVar or OccursVar instead - none of them preserve
occurrence information, and so it does not make sense for them to
operate on occurs vars.