Relax the value restriction for constrained values
Rationale: The value restriction is there to prevent us from treating
polymorphic /values/ as if they were functions of their type. E.g., in

    let foo : forall 'a. ref (option 'a) = ref None

Semantically, we're trying to treat foo as if it really took `'a` as a
parameter! (In the compiler's metatheory it does, but then types are
promptly erased for codegen.)

However, polymorphic values with constraints really are a function of
their constraints. This:

    let foo : forall 'a. eq 'a => ref (option 'a) = ref None

Means no harm.