Recheck files when their loading is interrupted
If a file was checked, and then interrupted due to a file change
downstream, then we would never resume checking this file. Instead, we'd
assume it had errored loading and thus abort.

This keeps track of when we last started working on a file, and thus
whether it needs to be reloaded or not.