Add basic support for "intrinsics"
This is effectively a more lightweight version of builtin
variables. One may declare an external value as follows:

  external val ( + ) : int -> int -> int = "%int.add"

This will now be special cased by the compiler, allowing us to do
constant folding and whatnot.

This commit doesn't add any optimisations at all, it just adds frontend
and backend support for intrinsics. For the sake of convenience, these
are assumed to always be binary operations